+++ example/ios/Runner/AppDelegate.swift
@@ -1,19 +1,16 @@
-import UIKit
 import Flutter
-import mParticle_Apple_SDK
+import UIKit
 
-@UIApplicationMain
-@objc class AppDelegate: FlutterAppDelegate {
+@main
+@objc class AppDelegate: FlutterAppDelegate, FlutterImplicitEngineDelegate {
   override func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
-    GeneratedPluginRegistrant.register(with: self)
-
-    let options = MParticleOptions(key: "api-key", secret: "secret")
-    options.logLevel = MPILogLevel.verbose
-    MParticle.sharedInstance().start(with: options)
-
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
+  }
+
+  func didInitializeImplicitFlutterEngine(_ engineBridge: FlutterImplicitEngineBridge) {
+    GeneratedPluginRegistrant.register(with: engineBridge.pluginRegistry)
   }
 }